Fjernet ubrukt task og dto, flyttet logikk knyttet til henleggelsesbrev til brevservice
