+++ src/main/kotlin/no/nav/familie/klage/brev/BrevClient.kt
@@ -7,7 +7,6 @@ import no.nav.familie.klage.brev.dto.FritekstBrevRequestDto
 import no.nav.familie.klage.felles.util.TekstUtil.norskFormat
 import no.nav.familie.klage.felles.util.medContentTypeJsonUTF8
 import no.nav.familie.klage.infrastruktur.exception.feilHvis
-import no.nav.familie.kontrakter.ef.felles.FrittståendeBrevDto
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpHeaders
@@ -74,8 +73,8 @@ class BrevClient(
         )
     }
 
-    fun sendFrittståendeBrev(frittståendeBrevDto: FrittståendeBrevDto) {
-        postForEntity<Any>(URI.create("http://familie-ef-iverksett/api/brev/frittstaende"), frittståendeBrevDto)
+    companion object {
+        const val FRITEKST = "fritekst"
     }
 }
 
@@ -92,5 +91,3 @@ data class BrevRequest(
     val skjulBeslutterSignatur: Boolean,
     val dato: String,
 )
-
-const val FRITEKST = "fritekst"
